requirements.txt
attrs==20.2.0
boto3==1.16.10
botocore==1.19.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flask-cors==3.0.9
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
httplib2==0.18.1
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
oauth2client==4.1.3
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pyyaml==5.3.1
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
tenacity==6.2.0
urllib3==1.25.11
Werkzeug==1.0.1
zipp==3.4.0
